Create a shader based on Tiled Directional Flow
See
<http://www.rug.nl/science-and-society/centre-for-information-technology/research/hpcv/publications/watershader/>.

The waves00.png texture is taken from the ZIP archive available for
download above.

It’s license is stated as:

                    This software is Creditware:

you can do whatever you want with this shader except claiming rights
you may sell it, but you cannot prevent others from selling it, giving
it away or use it as they please.

Having said that, it would be nice if you gave me some credit for it,
when you use it.