Split between sync() and prepare(RenderContext&)
sync() is called once for the scenegraph, while prepare() is called for
each RenderContext, that is, for each Scene.

Thus, in prepare(), everything related to the viewpoint can be set up
and must be stored keyed with the RenderContext in use. sync() however
can be used to transmit data needed for any perspective.

The key difference between prepare() and render() is that the data lock
is not held while render() is called.